Update Gundog.Tracks.R
New function input add on: 'max.speed'

If a value is given for max.speed (e.g., max.speed = 3), then prior to the VPC procedure, speed values that exceed the supplied max.speed value are re-allocated as the max.speed value (in this case, 3 m/s). This is only really relevant for speeds derived by the DBA approach. Note this occurs pre-current integration (if ch & cs are supplied) and VPC so resultant speeds may still be over this threshold